Moved types into the right place
