+++ src/queries.ts
@@ -1,6 +1,6 @@
 import { processStorableObjects, uploadStorableObjects } from '@/src/storage';
-import type { Query, Results } from '@/src/types/query';
-import type { QueryHandlerOptions } from '@/src/types/utils';
+import type { Query } from '@/src/types/query';
+import type { QueryHandlerOptions, Results } from '@/src/types/utils';
 import { WRITE_QUERY_TYPES } from '@/src/utils/constants';
 import { runQueriesWithHooks } from '@/src/utils/data-hooks';
 import {
