+++ src/types/utils.ts
@@ -1,10 +1,9 @@
 import type { AsyncLocalStorage } from 'node:async_hooks';
 
-import type { Query } from '@/src/types/query';
 import type { StorableObjectValue } from '@/src/types/storage';
 import type { Hooks } from '@/src/utils/data-hooks';
 
-import type { Model } from '@ronin/compiler';
+import type { Model, Query } from '@ronin/compiler';
 import type { RONIN } from './codegen';
 
 export interface QueryHandlerOptions {
@@ -145,3 +144,37 @@ export interface QueryItem {
   query: Query;
   options?: Record<string, unknown>;
 }
+
+type QueryPaginationOptions = {
+  moreBefore?: (string | null) | undefined;
+  moreAfter?: (string | null) | undefined;
+};
+
+/**
+ * Utility type that merges extra properties `R` into each element of array-type
+ * properties in a type `T`; non-array properties are left unchanged.
+ */
+type BindToArray<T, R> = {
+  [K in keyof T]: T[K] extends Array<infer U> ? Array<U> & R : T[K];
+};
+
+/**
+ * Utility type used to type the results of a query.
+ *
+ * It unwraps the promised type if `T` is an array of Promises, adds `moreBefore?`
+ * and `moreAfter?` fields to an array's items if `T` is an array. Otherwise it
+ * wraps non-array and non-Promise types in an array.
+ */
+export type Results<T> = T extends never | Array<never>
+  ? T
+  : T extends []
+    ? []
+    : T extends [infer First, ...infer Rest]
+      ? Rest extends unknown
+        ? First extends Promise<infer U>
+          ? [U]
+          : BindToArray<[First, ...Rest], QueryPaginationOptions>
+        : Array<First>
+      : T extends Promise<infer U>
+        ? [U]
+        : [T];
